Wordle Game
Green: If a letter is exactly in the right place, its box is shown in green.
Yellow: If a letter is in the word but in the wrong place, its box is shown in yellow.
Gray: If a letter is not in the word at all, its box is shown in gray.